Renamed Project
Renamed the project and repository from HeatseekerHeaven-Utilities to CTN_DiscordTS-Template.
The reason for this is that this was already meant to be a template from the start but it has gone to a point that it needs its own repository and I would rather start with the template and then fork it to an actual bot then the other way around.