Update errorHandler.ts
patched an error that happened once the open brace "(" was not present in a stackLine. this happens only when the stack trace comes from a call that happened inside a wrapped call like: functionCall().then(() => {//here}).
To fix this, it now just assigns call with a string and puts whatever is left in the path.