Transformed data
Command data now takes in buttons and select menus as a class, this is to allow giving buttons and selects a name which you can call on when you want to retrieve it. The structure has changed a lot and there is a few extra steps to making a complete command but all together i think the dev experience should improve from this.